fixed controllers
